Bug 1622530 - Part 5: Convert TryNoteKind into enum class. r=jorendorff
Differential Revision: https://phabricator.services.mozilla.com/D66906

--HG--
extra : moz-landing-system : lando

+++ js/src/frontend/BytecodeSection.cpp
@@ -140,7 +140,8 @@ bool CGTryNoteList::append(TryNoteKind kind, uint32_t stackDepth,
   // Offsets are given relative to sections, but we only expect main-section
   // to have TryNotes. In finish() we will fixup base offset.
 
-  TryNote note(kind, stackDepth, start.toUint32(), (end - start).toUint32());
+  TryNote note(uint32_t(kind), stackDepth, start.toUint32(),
+               (end - start).toUint32());
 
   return list.append(note);
 }
